re organize output code into its own module and tests
