+++ pyswmm/simulation.py
@@ -9,29 +9,7 @@
 
 # Local imports
 from pyswmm.swmm5 import PySWMM, PYSWMMException
-from pyswmm.errors import IncompleteSimulation
 from pyswmm.toolkitapi import SimulationTime, SimulationUnits
-# Third party imports
-from swmm.toolkit import output, shared_enum
-from datetime import timedelta
-from julian import from_jd
-
-
-def output_exception_handler(func):
-    def inner_function(self, *args, **kwargs):
-        if self.sim and not self.sim._complete:
-            raise IncompleteSimulation('Cannot use output functions due to incomplete simulation.')
-
-    return inner_function
-
-
-def output_open_handler(func):
-    def inner_function(self, *arg, **kwargs):
-        print('hi')
-        if not self._sim.output.loaded:
-            self._sim.output.open()
-
-    return inner_function
 
 
 class Simulation(object):
@@ -87,18 +65,12 @@ class Simulation(object):
                  reportfile=None,
                  outputfile=None,
                  swmm_lib_path=None):
-
         self._model = PySWMM(inputfile, reportfile, outputfile, swmm_lib_path)
         self._model.swmm_open()
         self._isOpen = True
         self._advance_seconds = None
         self._isStarted = False
         self._terminate_request = False
-        self._complete = False
-
-        self.output = Output(self._model.binfile, self)
-        self._times = None
-
         self._callbacks = {
             "before_start": None,
             "before_step": None,
@@ -141,7 +113,6 @@ class Simulation(object):
             self._execute_callback(self.before_start())
             self._model.swmm_start(True)
             self._isStarted = True
-            self._complete = False
 
     def __next__(self):
         """Next"""
@@ -162,8 +133,6 @@ class Simulation(object):
         self._execute_callback(self.after_step())
         if time <= 0.0:
             self._execute_callback(self.before_end())
-            self._complete = True
-            self.output.open()
             raise StopIteration
         return self._model
 
@@ -181,8 +150,6 @@ class Simulation(object):
             self._isOpen = False
             # Execute Callback Hooks After Simulation Closes
             self._execute_callback(self.after_close())
-        if not self.output.delete_handle and (self.output.loaded or self.output.handle):
-            self.output.close()
 
     @staticmethod
     def _is_callback(callable_object):
@@ -674,189 +641,3 @@ class Simulation(object):
         dt = self.current_time - self.start_time
         total_time = self.end_time - self.start_time
         return float(dt.total_seconds()) / total_time.total_seconds()
-
-    @property
-    def times(self):
-        if self.output and not self._times:
-            self.output.open()
-            start_date_time = self.output.start_time
-            num_steps = self.output.num_period
-            report_step = self.output.report
-            self._times = [start_date_time + timedelta(report_step) * step for step in range(num_steps)]
-        return self._times
-
-
-class Output(object):
-    def __init__(self, binfile, sim=None):
-        self.sim = sim
-        self.binfile = binfile
-
-        self.handle = None
-        self.loaded = False
-        self.delete_handle = False
-        self.num_period = None
-        self.report = None
-        self.start_time = None
-
-    #@output_exception_handler
-    def open(self):
-        """
-        Open a binary file
-        """
-        print('hi')
-        if self.handle is None:
-            self.handle = output.init()
-
-        if not self.loaded:
-            self.loaded = True
-            print(self.binfile)
-            output.open(self.handle, self.binfile)
-            self.start_time = from_jd(output.get_start_date(self.handle) + 2415018.5)
-            self.report = output.get_times(self.handle, shared_enum.Time.REPORT_STEP)
-            self.num_period = output.get_times(self.handle, shared_enum.Time.NUM_PERIODS)
-
-    def __enter__(self):
-        self.open()
-
-        return self
-
-    def __exit__(self):
-        self.close()
-
-    @output_exception_handler
-    def close(self):
-        """
-        Close an opened binary file
-        """
-        if self.handle or self.loaded:
-            self.loaded = False
-            self.delete_handle = True
-            output.close(self.handle)
-
-    @output_exception_handler
-    @output_open_handler
-    def project_size(self):
-        return output.get_proj_size(self.handle)
-
-    @output_exception_handler
-    @output_open_handler
-    def unit(self):
-        return output.get_units(self.handle)
-
-    @output_exception_handler
-    @output_open_handler
-    def version(self):
-        return output.get_version(self.handle)
-
-    @output_exception_handler
-    @output_open_handler
-    def object_name(self, object_type, index):
-        return output.get_elem_name(self.handle, object_type, index)
-
-    @output_exception_handler
-    @output_open_handler
-    def subcatch_series(self, index, attribute, start_index=None, end_index=None):
-        if not start_index:
-            start_index = 0
-
-        if not end_index:
-            end_index = self.num_period
-
-        return output.get_subcatch_series(self.handle, index, attribute, start_index, end_index)
-
-    @output_exception_handler
-    @output_open_handler
-    def node_series(self, index, attribute, start_index=None, end_index=None):
-        if not start_index:
-            start_index = 0
-
-        if not end_index:
-            end_index = self.num_period
-
-        return output.get_node_series(self.handle, index, attribute, start_index, end_index)
-
-    @output_exception_handler
-    @output_open_handler
-    def link_series(self, index, attribute, start_index=None, end_index=None):
-        if not start_index:
-            start_index = 0
-
-        if not end_index:
-            end_index = self.num_period
-
-        return output.get_link_series(self.handle, index, attribute, start_index, end_index)
-
-    @output_exception_handler
-    @output_open_handler
-    def system_series(self, attribute, start_index=None, end_index=None):
-        if not start_index:
-            start_index = 0
-
-        if not end_index:
-            end_index = self.num_period
-
-        return output.get_system_series(self.handle, attribute, start_index, end_index)
-
-    @output_exception_handler
-    @output_open_handler
-    def subcatch_attribute(self, attribute, time_index=0):
-        if not time_index:
-            time_index = 0
-
-        return output.get_subcatch_attribute(self.handle, time_index, attribute)
-
-    @output_exception_handler
-    @output_open_handler
-    def node_attribute(self, attribute, time_index=0):
-        if not time_index:
-            time_index = 0
-
-        return output.get_node_attribute(self.handle, time_index, attribute)
-
-    @output_exception_handler
-    @output_open_handler
-    def link_attribute(self, attribute, time_index=0):
-        if not time_index:
-            time_index = 0
-
-        return output.get_link_attribute(self.handle, time_index, attribute)
-
-    @output_exception_handler
-    @output_open_handler
-    def system_attribute(self, attribute, time_index=0):
-        if not time_index:
-            time_index = 0
-
-        return output.get_system_attribute(self.handle, time_index, attribute)
-
-    @output_exception_handler
-    @output_open_handler
-    def subcatch_result(self, index, time_index=0):
-        if not time_index:
-            time_index = 0
-
-        return output.get_subcatch_result(self.handle, time_index, index)
-
-    @output_exception_handler
-    @output_open_handler
-    def node_result(self, index, time_index=0):
-        if not time_index:
-            time_index = 0
-
-        return output.get_node_result(self.handle, time_index, index)
-
-    @output_exception_handler
-    @output_open_handler
-    def link_result(self, index, time_index=0):
-        if not time_index:
-            time_index = 0
-
-        return output.get_link_result(self.handle, time_index, index)
-
-    @output_exception_handler
-    @output_open_handler
-    def system_result(self, index, time_index=0):
-        if not time_index:
-            time_index = 0
-
-        return output.get_system_result(self.handle, time_index, index)
